Make functions for everything
Now I hope the code is easier to read and more understandable
remove fuctions are to remove a character from regex.
Coinc means coincidence and the function with coinc depends of what type is regex - without special characters, with a special one in the beggining or with a special character somewhere else. I'm not sure this code works properly, but I think this is a good start